Reorganise background screen code
